feat(utils): add directive indentation normalizer
