+++ apps/campfire/src/components/Passage/Passage.tsx
@@ -7,8 +7,7 @@ import {
   remarkParagraphStyles
 } from '@campfire/utils/remarkStyles'
 import { createMarkdownProcessor } from '@campfire/utils/createMarkdownProcessor'
-import { scanDirectives } from '@campfire/utils/scanDirectives'
-import { shouldStripDirectiveIndent } from '@campfire/utils/shouldStripDirectiveIndent'
+import { normalizeDirectiveIndentation } from '@campfire/utils/normalizeDirectiveIndentation'
 import {
   isTitleOverridden,
   clearTitleOverride
@@ -20,34 +19,6 @@ import {
 import { useDeckStore } from '@campfire/state/useDeckStore'
 import { componentMap } from '@campfire/components/Passage/componentMap'
 import type { WorkerRequest, WorkerResponse } from './directiveWorker'
-
-/**
- * Normalizes directive indentation so Markdown treats directive lines the same
- * regardless of leading spaces or tabs. Uses {@link scanDirectives} to walk the
- * source once and remove tabs or four-or-more spaces before directive markers.
- *
- * @param input - Raw passage text.
- * @returns Passage text with directive indentation normalized.
- */
-const normalizeDirectiveIndentation = (input: string): string => {
-  let output = ''
-  let lineStart = 0
-  for (const token of scanDirectives(input)) {
-    if (token.type === 'text') {
-      output += token.value
-    } else {
-      const indent = output.slice(lineStart).match(/^[\t ]*/)?.[0] ?? ''
-      if (shouldStripDirectiveIndent(indent))
-        output = output.slice(0, lineStart)
-      output += token.value
-    }
-    const lastNewline = token.value.lastIndexOf('\n')
-    if (lastNewline !== -1) {
-      lineStart = output.length - (token.value.length - lastNewline - 1)
-    }
-  }
-  return output
-}
 
 /**
  * Builds a document title from story and passage names.
