+++ apps/campfire/src/utils/__tests__/normalizeDirectiveIndentation.test.ts
@@ -1,0 +1,16 @@
+import { describe, it, expect } from 'bun:test'
+import { normalizeDirectiveIndentation } from '@campfire/utils/normalizeDirectiveIndentation'
+
+describe('normalizeDirectiveIndentation', () => {
+  it('strips tabs and four spaces before directives', () => {
+    const input = '\t:tab\n    :spaces\n  :keep\n:root'
+    const output = normalizeDirectiveIndentation(input)
+    expect(output).toBe(':tab\n:spaces\n  :keep\n:root')
+  })
+
+  it('keeps short or mixed indentation', () => {
+    const input = '   :three\n \t:mix'
+    const output = normalizeDirectiveIndentation(input)
+    expect(output).toBe('   :three\n \t:mix')
+  })
+})
